refactored commands to objects
