+++ test/commands/update-test.js
@@ -1,13 +1,13 @@
 /* global describe, afterEach, it */
 var expect = require('chai').expect
 
-var CommandTester = require('./lib/command-tester')
-var packageFixture = require('./lib/package-fixture-manager')
+var CommandTester = require('../lib/command-tester')
+var packageFixture = require('../lib/package-fixture-manager')
 
-var UpdateCommand = require('../commands/update')
+var UpdateCommand = require('../../commands/update')
 var sharedTester = new CommandTester(new UpdateCommand(), [])
 
-process.chdir(__dirname)
+process.chdir(__dirname + '/..')
 
 describe('update', function () {
   afterEach(function () {
