+++ test/integration/init-test.js
@@ -1,10 +1,10 @@
 /* global describe, it, beforeEach, afterEach */
 var exec = require('child_process').exec
 var expect = require('chai').expect
-var validAddress = require('./fixtures/test-address')
-var packageFixture = require('./lib/package-fixture-manager')
+var validAddress = require('../fixtures/test-address')
+var packageFixture = require('../lib/package-fixture-manager')
 
-process.chdir(__dirname)
+process.chdir(__dirname + '/..')
 
 describe('init [integration]', function () {
   beforeEach(function () { packageFixture.setup('empty')})
